Macro dxfilter::compile_shader

compile_shader!() { /* proc-macro */ }

Compile directx shader at compile time and returns byte code. it uses D3DCompile2. any unexplained parameters are analogous to that function.

Syntax

compile_shader!{
    src: "some shader source code"  // [required] either src or src_file is required
    src_file: "path/to/shader/source/code" // [required] either src or src_file is required
    entry_point: "main_func_name"   // [required] name of entry point function
    target: "shader_profile"        // [required] shader used to compile the shader.
    src_name: "path/to/shader/source/file"  // [optional] required for #include if any.
                                            // This is auto generated when src_file is used.
    defines: {                      // [optional] used to define shader macros before compiling
        ("DEFINE_1","32"),
    },
    flags1: 0,                      // [optional] flags1
    flags2: 0,                      // [optional] flags2
    secondary_data_flags: 0         // [optional] secondary_data_flags
    secondary_data: ""              // [optional] secondary_data
}

Example usage

let data = compile_shader!{
   src: "
        int main() {
            return 1;
        }
    ",
    entry_point: "main"
    target: "ps_5_0"
};

Compile directx shader at compile time and returns byte code. it uses D3DCompile2. any unexplained parameters are analogous to that function.

Syntax

compile_shader!{
    src: "some shader source code"  // [required] either src or src_file is required
    src_file: "path/to/shader/source/code" // [required] either src or src_file is required
    entry_point: "main_func_name"   // [required] name of entry point function
    target: "shader_profile"        // [required] shader used to compile the shader.
    src_name: "path/to/shader/source/file"  // [optional] required for #include if any.
                                            // This is auto generated when src_file is used.
    defines: {                      // [optional] used to define shader macros before compiling
        ("DEFINE_1","32"),
    },
    flags1: 0,                      // [optional] flags1
    flags2: 0,                      // [optional] flags2
    secondary_data_flags: 0         // [optional] secondary_data_flags
    secondary_data: ""              // [optional] secondary_data
}

Example usage

let data = compile_shader!{
   src: "
        int main() {
            return 1;
        }
    ",
    entry_point: "main"
    target: "ps_5_0"
};