Trait dxfilter::DxFilter

pub trait DxFilter {
    fn apply_filter(&self, ctx: &ID3D11DeviceContext4) -> Result<()>;
    fn set_input_tex(&mut self, tex: &Texture) -> Result<()>;
    fn set_output_tex(&mut self, tex: &Texture) -> Result<()>;
}

Interface for interacting with various filters. Interface is defined so that you could create Directx pipelines that involve multiple filters.

Example Usage:

fn main () {
    //...
    let (device, context) = // acquire Device and DeviceContext

    let input_tex = // create an input texture
    let output_tex = // create an output texture

    // create some directx filter. for example the following
    let filter = ScaleARGBOrAYUV::new(&input_tex,&output_tex,&device);

    // apply the filter
    filter.apply_filter(&context);

    // read from the output_texture
}

Required Methods

fn apply_filter(&self, ctx: &ID3D11DeviceContext4) -> Result<()>

takes directx device context and applies various vertex and pixel shaders to apply the filter.

fn set_input_tex(&mut self, tex: &Texture) -> Result<()>

configure the filter to use different input texture.

fn set_output_tex(&mut self, tex: &Texture) -> Result<()>

configure the filter to use different output texture.